Struct objc_encode::parse::StrEncoding

pub struct StrEncoding<S: ?Sized = str>(_)
 where
    S: AsRef<str>;

An Objective-C type encoding parsed from its string representation.

Encodings are parsed lazily; for example, when comparing for equality we may then find that the encoding is of a struct, but no further evaluation is done of the fields of the struct until they are requested.

Methods

impl StrEncoding

pub fn from_str(s: &str) -> Result<&StrEncoding, ParseEncodingError<&str>>

Constructs a StrEncoding parsed from the given string, or returns an error if the string was not a valid encoding.

from_str will return a reference to a dynamically-sized StrEncoding which is valid for the lifetime of the given string. To construct a sized StrEncoding that owns a buffer, use the new method.

pub fn from_str_unchecked(s: &str) -> &StrEncoding

Constructs a StrEncoding without checking to see if the given string is a valid encoding.

If the given string is not a valid encoding, the returned StrEncoding will panic later when being evaluated.

impl<S> StrEncoding<S> where
S: AsRef<str>,

pub fn new(s: S) -> Result<StrEncoding<S>, ParseEncodingError<S>>

Constructs a StrEncoding parsed from the given string, or returns an error if the string was not a valid encoding.

new constructs a sized StrEncoding that owns the given buffer, meaning it can be moved without dealing with lifetimes. If you only have a borrowed &str slice or don't want to allocate a buffer, use the from_str method.

pub fn new_unchecked(s: S) -> StrEncoding<S>

Constructs a StrEncoding without checking to see if the given string is a valid encoding.

If the given string is not a valid encoding, the returned StrEncoding will panic later when being evaluated.

impl<S: ?Sized> StrEncoding<S> where
S: AsRef<str>,

pub fn as_str(&self) -> &str

Returns the string representation of self.