Trait objc_encode::EncodingsIterateCallback

pub trait EncodingsIterateCallback {
    fn call<T: ?Sized + Encoding>(&mut self, _: &T) -> bool;
}

Types that can be used as callbacks while iterating over Encodings.

A special trait is necessary rather than a normal closure because the type of each Encoding may be different.

Required Methods

fn call<T: ?Sized + Encoding>(&mut self, _: &T) -> bool

This method is called for each encoding of an Encodings. The return value indicates whether iteration should stop; return true to stop the iteration before completion.