Trait objc_encode::Encodings

pub trait Encodings {
    fn each<F: EncodingsIterateCallback>(&self, &mut F);
    fn eq_encodings<T: ?Sized + Encodings>(&self, encs: &T) -> bool;

    fn write_all<W: Write>(&self, writer: &mut W) -> Result { ... }
}

Types that represent a collection of Encodings.

Required Methods

fn each<F: EncodingsIterateCallback>(&self, &mut F)

Iterates over the encodings of Self, calling the provided callback for each.

fn eq_encodings<T: ?Sized + Encodings>(&self, encs: &T) -> bool

Returns whether Self is equal to the given encodings,

Provided Methods

fn write_all<W: Write>(&self, writer: &mut W) -> Result

Writes each encoding of Self in order to the given writer.